More functions/macros moved to private/Rinternals.h
