+++ src/appl/lbfgsb.c
@@ -35,7 +35,7 @@
 #include <float.h> /* for DBL_EPSILON */
 #include <string.h>
 #include <R_ext/RS.h> /* for F77_CALL */
-#include <R_ext/Linpack.h>
+#include <R_ext/private/Linpack.h>
 #include <R_ext/Applic.h>
 #include <R_ext/Print.h> /* Rprintf */
