+++ src/include/R_ext/Utils.h
@@ -40,8 +40,10 @@
 # define R_SIZE_T size_t
 #endif
 
-#define revsort       Rf_revsort
-#define iPsort        Rf_iPsort
+//@MOVED: Rf_revsort unused
+//#define revsort       Rf_revsort
+//@MOVED: Rf_iPsort unused
+//#define iPsort        Rf_iPsort
 #define rPsort        Rf_rPsort
 //@MOVED: cPsort used 1 times in main
 //#define cPsort        Rf_cPsort
@@ -61,12 +63,16 @@ extern "C" {
 /* ../../main/sort.c : */
 void	R_isort(int*, int);
 void	R_rsort(double*, int);
-void	R_csort(Rcomplex*, int);
+//@MOVED: R_csort used 1 times in main
+//void	R_csort(Rcomplex*, int);
 void    rsort_with_index(double *, int *, int);
-void	revsort(double*, int*, int);/* reverse; sort i[] alongside */
-void	iPsort(int*,    int, int);
+//@MOVED: Rf_revsort unused
+//void	revsort(double*, int*, int);/* reverse; sort i[] alongside */
+//@MOVED:
+//void	iPsort(int*,    int, int);
 void	rPsort(double*, int, int);
-void	cPsort(Rcomplex*, int, int);
+//@MOVED: cPsort used 1 times in main
+//void	cPsort(Rcomplex*, int, int);
 
 /* ../../main/qsort.c : */
 /* dummy renamed to II to avoid problems with g++ on Solaris */
@@ -74,18 +80,26 @@ void R_qsort    (double *v,         R_SIZE_T i, R_SIZE_T j);
 void R_qsort_I  (double *v, int *II, int i, int j);
 void R_qsort_int  (int *iv,         R_SIZE_T i, R_SIZE_T j);
 void R_qsort_int_I(int *iv, int *II, int i, int j);
-#ifdef R_RS_H
-void F77_NAME(qsort4)(double *v, int *indx, int *ii, int *jj);
-void F77_NAME(qsort3)(double *v,            int *ii, int *jj);
-#endif
+//@MOVED:
+//#ifdef R_RS_H
+//@MOVED: qsort4_ unused
+//void F77_NAME(qsort4)(double *v, int *indx, int *ii, int *jj);
+//@MOVED: qsort3_ unused
+//void F77_NAME(qsort3)(double *v,            int *ii, int *jj);
+//#endif
 
 /* ../../main/util.c  and others : */
 const char *R_ExpandFileName(const char *);
-void	setIVector(int*, int, int);
-void	setRVector(double*, int, double);
-Rboolean StringFalse(const char *);
-Rboolean StringTrue(const char *);
-Rboolean isBlankString(const char *);
+//@MOVED: setIVector used 1 times in main
+//void	setIVector(int*, int, int);
+//@MOVED: Rf_setRVector unused
+//void	setRVector(double*, int, double);
+//@MOVED: Rf_StringFalse unused
+//Rboolean StringFalse(const char *);
+//@MOVED: Rf_StringTrue unused
+//Rboolean StringTrue(const char *);
+//@MOVED: Rf_isBlankString unused
+//Rboolean isBlankString(const char *);
 
 /* These two are guaranteed to use '.' as the decimal point,
    and to accept "NA".
@@ -94,7 +108,8 @@ double R_atof(const char *str);
 double R_strtod(const char *c, char **end);
 
 char *R_tmpnam(const char *prefix, const char *tempdir);
-char *R_tmpnam2(const char *prefix, const char *tempdir, const char *fileext);
+//@MOVED: R_tmpnam2 used 6 times in main
+//char *R_tmpnam2(const char *prefix, const char *tempdir, const char *fileext);
 
 void R_CheckUserInterrupt(void);
 void R_CheckStack(void);
@@ -105,16 +120,18 @@ void R_CheckStack2(R_SIZE_T);
 int findInterval(double *xt, int n, double x,
 		 Rboolean rightmost_closed,  Rboolean all_inside, int ilo,
 		 int *mflag);
-int findInterval2(double *xt, int n, double x,
-		  Rboolean rightmost_closed,  Rboolean all_inside, Rboolean left_open,
-		  int ilo, int *mflag);
+//@MOVED: findInterval2 used 5 times in appl, main
+//int findInterval2(double *xt, int n, double x,
+//		  Rboolean rightmost_closed,  Rboolean all_inside, Rboolean left_open,
+//		  int ilo, int *mflag);
 #ifdef R_RS_H
 int F77_SUB(interv)(double *xt, int *n, double *x,
 		    Rboolean *rightmost_closed, Rboolean *all_inside,
 		    int *ilo, int *mflag);
 #endif
-void find_interv_vec(double *xt, int *n,	double *x,   int *nx,
-		     int *rightmost_closed, int *all_inside, int *indx);
+//@MOVED: find_interv_vec unused
+//void find_interv_vec(double *xt, int *n,	double *x,   int *nx,
+//		     int *rightmost_closed, int *all_inside, int *indx);
 
 /* ../../appl/maxcol.c: also in Applic.h */
 void R_max_col(double *matrix, int *nr, int *nc, int *maxes, int *ties_meth);
