+++ src/library/stats/src/rWishart.c
@@ -24,7 +24,7 @@
 #include <math.h>
 #include <string.h>  // memset, memcpy
 #include <R.h>
-#include <Rinternals.h>
+#include <private/Rinternals.h>
 #include <Rmath.h>
 #include <R_ext/Lapack.h>        /* for Lapack (dpotrf, etc.) and BLAS */
