+++ src/modules/internet/sockconn.c
@@ -29,7 +29,7 @@
 #define R_USE_SIGNALS 1
 #include <Defn.h>
 #include <Rconnections.h>
-#include <R_ext/R-ftp-http.h>
+#include <R_ext/private/R-ftp-http.h>
 #include "sock.h"
 #include <errno.h>
